requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==1.4.1
Pillow==9.5.0
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
torch==1.13.1
torchvision==0.5.0
tqdm==4.65.0
